requirements.txt
requests
python-dotenv
matplotlib
numpy
json2html